requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
discord-typings==0.7.0
emoji==2.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.4
orjson==3.9.10
py-cord==2.4.1
pycares==4.4.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
roblox==2.0.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
